# docs/swagger.json
{
	"components": {
...
		"schemas": {
			"ICalculator": {
				"properties": {
					"user": {
						"type": "string"
					},
					"expression": {
						"type": "string"
					},
					"result": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"expression",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "calculator_server",
		"version": "1.0.0",
		"description": "A simple server for a calculator app written in typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/calculator/save": {
			"post": {
				"operationId": "Save",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"userId": "6142b4a3b0d7f4f8a7f8b7f6",
											"expression": "1+1",
											"result": 2
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "calculation saved successfully",
											"data": [
												{
													"expression": "1+1",
													"result": "2",
													"timestamp": "2021-09-15T17:07:15.000Z",
													"id": "6142b4a3b0d7f4f8a7f8b7f6"
												}
											],
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "calculation saved successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"CALCULATOR"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICalculator"
							}
						}
					}
				}
			}
		},
		"/calculator/get-history": {
			"get": {
				"operationId": "GetHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "calculation saved successfully",
											"data": [
												{
													"expression": "1+1",
													"result": "2",
													"timestamp": "2021-09-15T17:07:15.000Z",
													"id": "6142b4a3b0d7f4f8a7f8b7f6"
												}
											],
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "history fetched successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"CALCULATOR"
				],
				"security": [],
				"parameters": []
			}
		},
		"/calculator/get-history-by-user": {
			"get": {
				"operationId": "GetHistoryByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"userId": "6142b4a3b0d7f4f8a7f8b7f6"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "calculation saved successfully",
											"data": [
												{
													"expression": "1+1",
													"result": "2",
													"timestamp": "2021-09-15T17:07:15.000Z",
													"id": "6142b4a3b0d7f4f8a7f8b7f6"
												}
											],
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "history fetched successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"CALCULATOR"
				],
				"security": [],
				"parameters": [
					{
						"description": "- user id",
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/calculator/get-some-history": {
			"get": {
				"operationId": "GetSomeHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"limit": 10,
											"skip": 0
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "calculation saved successfully",
											"data": [
												{
													"expression": "1+1",
													"result": "2",
													"timestamp": "2021-09-15T17:07:15.000Z",
													"id": "6142b4a3b0d7f4f8a7f8b7f6"
												}
											],
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "history fetched successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"CALCULATOR"
				],
				"security": [],
				"parameters": [
					{
						"description": "- number of calculations to fetch",
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- number of calculations to skip",
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}